refactor: surplus MATIC test
The test was failing because it wasn't sending enough MATIC to OffsetHelper.

I started using `howMuchETHShouldISendToSwap` in the test which fixed it, but I realised the error message from Sushi is not useful.

So I implemented a check with a useful message within the OffsetHelper contract for when someone doesn't send enough ETH.