fix: calculate token amount method
While working on PRT-198 I realised we need a method to calculate the token amount needed to swap to NCT and redeem+retire in the same manner we have the howMuchETHShouldISendToSwap method that calculates the amount of ETH to send.

While working on this PR (PRT-212) I realised there was an issue with the swap.

Because there was not method to calculate the needed tokens, the code could break since approval and SafeTransferFrom in the swap method was done blindly.

It really only passed tests because I was using weth so the approved amount was always larger than needed, but it could have created cases were bugs happened.

P.S.: I have opened PRT-214 to refactor this as it's quite messy, but I am focused on the SDK right now.

P.P.S.: I have opened PRT-215 to write USDC and WMATIC tests.